+++ src/main/java/ru/yandex/practicum/filmorate/ErrorsIO/ValidationException.java
@@ -1,11 +1,9 @@
 package ru.yandex.practicum.filmorate.ErrorsIO;
 
 import lombok.Data;
-import lombok.extern.slf4j.Slf4j;
 
-@Slf4j
 @Data
-public class ValidationException extends Throwable {
+public class ValidationException extends RuntimeException {
     private String s;
 
     public ValidationException(String s) {
