+++ src/main/java/ru/yandex/practicum/filmorate/Servis/ServisDate.java
@@ -1,0 +1,7 @@
+package ru.yandex.practicum.filmorate.Servis;
+
+import java.time.LocalDateTime;
+
+public class ServisDate {
+    public static final LocalDateTime localDateTimeMinFilm = LocalDateTime.of(1895, 12, 28, 0, 0);
+}
